# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/teachers": {
      "get": {
        "operationId": "TeacherController_getTeachers",
        "parameters": [
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teachers have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all teachers",
        "tags": [
          "Teachers"
        ]
      },
      "post": {
        "operationId": "TeacherController_createTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teacher has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new teacher",
        "tags": [
          "Teachers"
        ]
      }
    },
    "/teachers/{id}": {
      "patch": {
        "operationId": "TeacherController_updateTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Teacher has been successfully updated."
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a teacher",
        "tags": [
          "Teachers"
        ]
      },
      "delete": {
        "operationId": "TeacherController_deleteTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Teacher has been successfully deleted."
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a teacher",
        "tags": [
          "Teachers"
        ]
      }
    },
...
    "/telegram/generate-telegram-link": {
      "post": {
        "operationId": "TelegramController_generateTelegramLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramLinkInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Telegram link has been successfully generated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLinkOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a telegram link for a teacher",
        "tags": [
          "Telegram"
        ]
      }
    },
    "/telegram/send-lessons-cost-to-admin": {
      "post": {
        "operationId": "TelegramController_sendLessonsCostToAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonsCostFiltersDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Lessons cost has been successfully sent to admin."
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send lessons cost to admin",
        "tags": [
          "Telegram"
        ]
      }
    }
  },
  "info": {
    "title": "School API",
    "description": "The School API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BadRequestErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "The HTTP status code",
            "example": 400
          },
          "path": {
            "type": "string",
            "description": "The path of the request",
            "example": "/api/v1/users"
          },
          "message": {
            "type": "object",
            "description": "The message of the error",
            "example": [
              "Error message 1",
              "Error message 2"
            ]
          }
        },
        "required": [
          "statusCode",
          "path",
          "message"
        ]
      },
...
      "TeacherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The id of the teacher",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the teacher",
            "example": "John Doe"
          },
          "login": {
            "type": "string",
            "description": "The login of the teacher",
            "example": "john.doe"
          },
          "role": {
            "type": "string",
            "description": "The role of the teacher",
            "example": "TEACHER"
          },
          "timezone": {
            "type": "string",
            "description": "The timezone of the teacher",
            "example": "BY"
          },
          "deleted_at": {
            "type": "object",
            "description": "The deleted at of the teacher",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "The created at of the teacher",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "telegrams": {
            "description": "The telegrams of the teacher",
            "example": [
              {
                "id": 1,
                "telegram_id": "1234567890",
                "username": "john.doe",
                "first_name": "John",
                "type": "STUDENT"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "login",
          "role",
          "timezone",
          "deleted_at",
          "created_at",
          "telegrams"
        ]
      },
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "The login of the teacher",
            "example": "teacher123"
          },
          "password": {
            "type": "string",
            "description": "The password of the teacher",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "description": "The name of the teacher",
            "example": "John Doe"
          },
          "timezone": {
            "type": "string",
            "description": "The timezone of the teacher",
            "example": "BY"
          }
        },
        "required": [
          "login",
          "password",
          "name",
          "timezone"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the teacher",
            "example": "John Doe"
          },
          "timezone": {
            "type": "string",
            "description": "The timezone of the teacher",
            "example": "BY"
          }
        },
        "required": [
          "name",
          "timezone"
        ]
      },
...
      "TelegramLinkInputDto": {
        "type": "object",
        "properties": {
          "teacher_id": {
            "type": "object",
            "description": "The teacher id",
            "example": 1
          },
          "student_id": {
            "type": "object",
            "description": "The student id",
            "example": 1
          }
        },
        "required": [
          "teacher_id",
          "student_id"
        ]
      },
      "TelegramLinkOutputDto": {
        "type": "object",
        "properties": {
          "link": {
            "type": "string",
            "description": "The link to the telegram bot",
            "example": "https://t.me/teacher_bot?start=1234567890"
          }
        },
        "required": [
          "link"
        ]
      },
      "LessonsCostFiltersDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "description": "The student id",
            "example": 1
          },
          "start_date": {
            "type": "string",
            "description": "The start date",
            "example": "2024-01-01"
          },
          "end_date": {
            "type": "string",
            "description": "The end date",
            "example": "2024-01-01"
          }
        },
        "required": [
          "student_id",
          "start_date",
          "end_date"
        ]
      }
    }
  }
}